+++ src/pagan_lab_to_nwb/interfaces/__init__.py
@@ -1,1 +1,1 @@
-from .behaviorinterface import BControlBehaviorInterface
+from .bcontroldatainterface import BControlBehaviorInterface
